requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
PyPDF2>=3.0.0
pandas>=2.0.0
pdf2image>=1.16.0 